% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa.R, R/extract_meta.R
\name{APA}
\alias{APA}
\alias{APA.psych}
\alias{APA.psych_alpha}
\alias{APA.coxph}
\alias{APA.summary.table}
\alias{APA.survdiff}
\alias{APA.NULL}
\alias{APA.default}
\alias{APA.QuadTypeIndependenceTest}
\alias{APA.ScalarIndependenceTest}
\alias{APA.numeric}
\alias{APA.formula}
\alias{APA.lm}
\alias{APA.glm}
\alias{APA.htest}
\alias{APA.xtabs}
\alias{APA.table}
\alias{APA.meta}
\alias{APA.metabias}
\title{Text-Ausgabe von Tests}
\usage{
APA(x, ...)

\method{APA}{psych}(x, ...)

\method{APA}{psych_alpha}(x, ...)

\method{APA}{coxph}(x, ...)

\method{APA}{summary.table}(x, ...)

\method{APA}{survdiff}(x, ...)

\method{APA}{`NULL`}(x, ...)

\method{APA}{default}(x, ...)

\method{APA}{QuadTypeIndependenceTest}(x, ...)

\method{APA}{ScalarIndependenceTest}(x, ...)

\method{APA}{numeric}(x, ...)

\method{APA}{formula}(x, data, exclude = NA, max_factor_length = 25, max_value = FALSE, ...)

\method{APA}{lm}(x, include.r = TRUE, ...)

\method{APA}{glm}(x, ...)

\method{APA}{htest}(x, ...)

\method{APA}{xtabs}(x, ...)

\method{APA}{table}(x, ...)

\method{APA}{summary.table}(x, ...)

\method{APA}{meta}(x, ...)

\method{APA}{metabias}(x)
}
\arguments{
\item{x}{Objekt fit, formula usw}

\item{...}{digits, data, usw}

\item{data, exclude, max_factor_length}{bei der Verwendung von Formeln}

\item{include.r}{APA.lm: R-Squar}
}
\value{
Character Vector mit einem oder meheren Eintraegen
}
\description{
Ausgabe von APA-Style formatiertem Text.
}
\examples{

  \donttest{
APA(mpg ~ cyl, mtcars)
APA(glm(vs ~ mpg, mtcars, family = binomial()))
APA(lm(mpg ~ drat + wt + qsec, mtcars))
APA(aov(mpg ~ drat + wt + qsec, mtcars))
APA(wilcox.test(mpg ~ vs, mtcars))
APA(t.test(mpg ~ vs, mtcars))

a <- letters[1:3]
APA(summary(table(a, sample(a))))

TeaTasting <-
  matrix(c(3, 1, 1, 3),
         nrow = 2,
         dimnames = list(
           Guess = c("Milk", "Tea"),
           Truth = c("Milk", "Tea")
         ))
APA(fisher.test(TeaTasting, alternative = "greater"))

# Test Survival Curve Differences
require(survival)
APA(survdiff(Surv(futime, fustat) ~ ecog.ps, data = ovarian))



require(coin)
davis <- matrix(c(3, 6, 2, 19),
                nrow = 2, byrow = TRUE)
davis <- as.table(davis)
## Asymptotic Pearson chi-squared test
diffusion <- data.frame(
  pd = c(
    0.80,    0.83,    1.89,    1.04,    1.45,
    1.38,    1.91,    1.64,    0.73,    1.46,
    1.15,    0.88,    0.90,    0.74,    1.21),
  age = factor(rep(c( "At term", "12-26 Weeks"), c(10, 5)))
)

## Exact Wilcoxon-Mann-Whitney test
## Hollander and Wolfe (1999, p. 111)
## (At term - 12-26 Weeks)
wt <- wilcox_test(pd ~ age,
                  data = diffusion,
                  distribution = "exact",
                  conf.int = TRUE)

APA(xt <- chisq_test(davis))
APA(wt)
}

require(meta)
data(Olkin1995)
m1 <- metabin(ev.exp, n.exp, ev.cont, n.cont,
              data = Olkin1995, subset = c(41, 47, 51, 59),
              studlab = paste(author, year),
              sm = "RR", method = "I")

APA(m1)

# Standard funnel plot
# sensitivity analysis, and reporting of Egger’s test
metabias(m1, method.bias = "linreg") |> APA()
}
