% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll_ranking.R
\name{cleanup_Rank}
\alias{cleanup_Rank}
\title{cleanup_Rank}
\usage{
cleanup_Rank(x, col.names = names(x))
}
\arguments{
\item{x}{data.frame}
}
\value{
data.frame
}
\description{
Doppelte Einträge bereinigen
}
\examples{

 dat <-  data.frame(
a = c(NA, "c", "a", NA,   "a", "a", "b", "a"),
b = c("c", NA, "b", NA,  "a", "a", "b", "b"),
c = c("a", "a", "b", NA,   NA, "a", "b", "c"),
d = c(NA, NA, NA, NA,     "a", "a", NA, "d"),
e = c(NA, "e", "a", NA,   "d",   NA, NA, "e")
)
cleanup_Rank(dat)
}
