% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll-roc.R
\name{extract.roc}
\alias{extract.roc}
\title{Youden’s J statistic}
\usage{
extract.roc(x, ..., digits = 1)
}
\description{
Optimal (Youden Index) point
}
\details{
The Youden index uses the maximum vertical distance of the ROC curve from the point (X, Y)
on the diagonal (random line). In fact, the Youden index maximizes the difference
between the Se and FP rate, in other words, it maximizes the percentage of Net correct
classification:

Therefore, the optimal cut-off point is calculated by maximizing Se+Sp at
different cut-off points

Quelle: https://bmcmedresmethodol.biomedcentral.com/articles/10.1186/s12874-024-02198-2
}
