% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Info_Statistic}
\alias{Info_Statistic}
\title{Information ueber die verwendeten Methoden}
\usage{
Info_Statistic(
  methode = "describe",
  library = "base",
  fun = "summary",
  my_methodes = ""
)
}
\arguments{
\item{methode, library, fun}{Text}
}
\value{
data.frame()
}
\description{
Information ueber die verwendeten Methoden
}
