% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{Mittelwert}
\alias{Mittelwert}
\alias{mean2}
\alias{sd2}
\alias{length2}
\alias{median2}
\alias{percent2}
\alias{calc_percent}
\alias{calc_mean}
\title{Mean}
\usage{
mean2(x, na.rm = TRUE)

sd2(x, na.rm = TRUE)

length2(x, na.rm = FALSE)

median2(x, na.rm = TRUE, ...)

percent2(x, ...)

calc_percent(
  x,
  digits = get_opt("prozent", "digits"),
  n = length(x),
  exclude = c(NA, NaN),
  max_factor_length = 25,
  style = get_opt("prozent", "style"),
  is_true_false = FALSE,
  ...
)

calc_mean(
  x,
  digits = get_opt("mean", "digits"),
  n = length(x),
  style = get_opt("mean", "style"),
  unit = NULL
)
}
\arguments{
\item{x}{vector}

\item{digits}{nachkomastellen (2)}

\item{n}{laenge des vectors}

\item{style}{lang oder Kurz}

\item{unit}{Einheiten}
}
\description{
Lagemasse Berechnen

Format Prozent
 style = 1    50% (27)
 style = 2    27 (50%)
 style = 3    50%
 style = 4    27
 style = 5    27/54
}
\examples{

mean2(rnorm(100))
sd2(rnorm(100))

percent2(gl(2, 8, labels = c("Control", "Treat")))
}
