% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-model_info.R
\name{model_info}
\alias{model_info}
\alias{model_info.default}
\alias{model_info.data.frame}
\alias{Infomation}
\alias{model_info.formula}
\alias{model_info.eff}
\alias{model_info.htest}
\alias{model_info.lm}
\alias{model_info.glm}
\alias{model_info.anova}
\alias{model_info.lmerModLmerTest}
\alias{model_info.merModLmerTest}
\alias{model_info.lmerTest}
\alias{model_info.glmerMod}
\alias{model_info.lme}
\alias{model_info.ScalarIndependenceTest}
\alias{model_info.biVar}
\alias{model_info.survfit}
\alias{model_info.coxph}
\alias{model_info.survdiff}
\alias{model_info.polr}
\title{model_info}
\usage{
model_info(x)

\method{model_info}{default}(x)

\method{model_info}{data.frame}(x)

Infomation(...)

\method{model_info}{formula}(x)

\method{model_info}{eff}(x)

\method{model_info}{htest}(x)

\method{model_info}{lm}(x)

\method{model_info}{glm}(x)

\method{model_info}{anova}(x)

\method{model_info}{lmerModLmerTest}(x)

\method{model_info}{merModLmerTest}(x)

\method{model_info}{lmerTest}(x)

\method{model_info}{glmerMod}(x)

\method{model_info}{lme}(x)

\method{model_info}{ScalarIndependenceTest}(x)

\method{model_info}{biVar}(x)

\method{model_info}{survfit}(x)

\method{model_info}{coxph}(x)

\method{model_info}{survdiff}(x)

\method{model_info}{polr}(x)
}
\arguments{
\item{x}{Fit Objekt}
}
\value{
Liste mit c("class",  "family", "y", "x", "labels", "N")
}
\description{
Extrahiert die Variablen namen die Labels und die Daten
}
\examples{
model_info(mpg ~ cyl)
model_info(glm(vs ~ mpg, mtcars, family = binomial()))
model_info(lm(mpg ~ drat + wt + qsec, mtcars))
model_info(wilcox.test(mpg ~ vs, mtcars))
model_info(t.test(mpg ~ vs, mtcars))

# model_info(Hmisc::spearman2(mpg ~ factor(vs), mtcars))

# model_info(coin::wilcox_test(mpg ~ factor(vs), mtcars))

# model_info(Hmisc::spearman2(mpg ~ factor(vs), mtcars))

}
