% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experimentelle-funktionen.R
\name{Post_Hoc}
\alias{Post_Hoc}
\title{pairwise.wilcox.test}
\usage{
Post_Hoc(
  ...,
  p.adjust.method = "bonf",
  test = "wilcox.test",
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  paired = FALSE,
  exact = NULL,
  correct = TRUE,
  conf.int = FALSE,
  conf.level = 0.95,
  tol.root = 1e-04,
  digits.rank = Inf
)
}
\arguments{
\item{...}{data, formula, names}

\item{p.adjust.method}{method for adjusting p value c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY",}

\item{alternative, mu, paired, exact, correct, conf.int, conf.level, tol.root, digits.rank}{not used}
}
\value{
data.frame
}
\description{
Post-hoc-Tests sind eine Familie von statistischen Tests, von denen es mehrere gibt.
}
\details{
Tukey HSD, wird verwendet, um alle Gruppen miteinander zu vergleichen (also alle möglichen Vergleiche von 2 Gruppen).

  Dunnett, um Vergleiche mit einer Referenzgruppe (Kontrollgruppe) durchzuführen.

  Bonferroni-Korrektur, wenn eine Reihe von geplanten Vergleichen durchgeführt werden soll.
}
\examples{
#' require(stp25stat2)
airquality$Month <- factor(airquality$Month)
Post_Hoc(airquality, Ozone, Month)
Post_Hoc(airquality, Ozone, Month, test= "t.test")
res_aov<-aov( Ozone ~ Month, airquality)
TukeyHSD( res_aov ) \%>\% Tbll()
DescTools::PostHocTest( res_aov, method = "hsd")
# Dunnett's test:
library(multcomp)
post_test <-
  glht(res_aov,
       linfct = mcp(Month = "Dunnett"))

summary(post_test)
summary(lm(res_aov))
}
