% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-squer.R
\name{R2}
\alias{R2}
\alias{R2.list}
\alias{R2.lm}
\alias{R2.coxph}
\alias{R2.glm}
\alias{R2.polr}
\alias{R2.mlm}
\alias{R2.merMod}
\alias{R2.lme}
\alias{RMSE}
\alias{RMSE.default}
\alias{RMSE.mlm}
\alias{RMSE.lmerModLmerTest}
\alias{RMSE.lmerMod}
\title{R-Quadrats}
\usage{
R2(x, ...)

\method{R2}{list}(...)

\method{R2}{lm}(x, ...)

\method{R2}{coxph}(x, ...)

\method{R2}{glm}(x, ...)

\method{R2}{polr}(x, ...)

\method{R2}{mlm}(x, ...)

\method{R2}{merMod}(x, ...)

\method{R2}{lme}(x, ...)

RMSE(x, ...)

\method{RMSE}{default}(x, ...)

\method{RMSE}{mlm}(x, ...)

\method{RMSE}{lmerModLmerTest}(x, ...)

\method{RMSE}{lmerMod}(x, ...)
}
\arguments{
\item{x}{fit-Objekt lm glm}

\item{...}{weitere Objekte nicht benutzt}
}
\value{
ein data.frame-Objekt.
}
\description{
Achtung es gibt noch die Funktion caret::R2 die Probleme macht

APA_R2: R2-Tabelle (noch nicht fretig)

coxph:
	an approximate R^2 based on Nagelkirke (Biometrika 1991)

glm:

McFadden:  McFadden's pseudo r-squared

 r2ML: Cox & Snell, Maximum likelihood pseudo r-squared

 r2CU: Nagelkerke Cragg and Uhler's pseudo r-squared

R2: The RMSE is the square root of the variance of the residuals.
Compute the root mean squared error
(see \code{sigma})

sigma: Residual standard error  RMSE:
 Root Mean Square Error
 RMSE.lmerModLmerTest sjstats::rmse(x)
}
\details{
Cox und Snell R2: [ 0.2 = akzeptabel, 0.4 = gut ]
 Nagelkerke R2: [ 0.2 = akzeptabel, 0.4 = gut, 0.5 = sehr gut]
 McFaddens R2: [ 0.2 = akzeptabel, 0.4 = gut ]
(see pscl::pR2)

Marginal and conditional r-squared for lme objects


For mixed-effects models, R2 can be categorized into two types. Marginal R2  represents the
variance explained by fixed factors

Conditional R2is interpreted as variance explained by both fixed and
random factors (i.e. the entire model).



 \code{MuMIn::r.squaredGLMM(x, ...)}
Pseudo-R-squared for Generalized Mixed-Effect models

For mixed-effects models, R² comes in two types: marginal and conditional.

Marginal R² represents the variance explained by the fixed effects.


  Conditional R² is interpreted as a variance explained by the entire model,
  including both fixed and random effects.


for R2.lme an lme model (usually fit using \code{lme}
This method extracts the variance for fixed and random effects,
as well as residuals, and calls \code{rsquared.glmm}

Marginal and conditional r-squared for merMod objects

This method extracts the variance for fixed and random effects, residuals,
and the fixed effects for the null model (in the case of Poisson family),
and calls \code{rsquared.glmm}

an merMod model (usually fit using lme4::lmer,
    lme4::glmer,lmerTest::lmer,
     blme::blmer, blme::bglmer, etc)

Marginal and conditional r-squared for glmm given fixed and random variances

This function is based on Nakagawa and Schielzeth (2013). It returns the marginal
and conditional r-squared, as well as the AIC for each glmm.
Users should call the higher-level generic "r.squared", or implement a method for the
corresponding class to get varF, varRand and the family from the specific object

 return A data frame with "Class", "Family", "Marginal", "Conditional", and "AIC" columns
}
\examples{
 \donttest{
 fit1<-lm(score ~ grade + treatment, schools)
 fit2<-lm(score ~ grade + treatment + stdTest, schools)
 R2.list(fit1, fit2)
}
 \donttest{
 lm1 <- lm(breaks ~ wool + tension, data = warpbreaks)
R2(lm1)
broom::glance(lm1)[c( "r.squared", "adj.r.squared")]
}
 \donttest{
lm2 <- glm(wool~ breaks + tension, warpbreaks, family= binomial())
R2(lm2)
#DescTools::PseudoR2(lm2,c("McFadden", "CoxSnell", "Nagelkerke"))
}
 \donttest{
 require(MASS)
options(contrasts = c("contr.treatment", "contr.poly"))
house.plr <- polr(Sat ~ Infl + Type + Cont, weights = Freq, data = housing)
R2(house.plr)
}

}
