% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{get_opt}
\alias{set_opt}
\alias{which_output}
\alias{farbe}
\alias{standard_theme}
\alias{ggplot_theme}
\alias{bw_theme}
\alias{wrap_string}
\alias{Cs}
\alias{contr.Treatment}
\alias{contr.Sum}
\alias{contr.Helmert}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Hmisc}{\code{\link[Hmisc]{Cs}}}

  \item{car}{\code{\link[car:Contrasts]{contr.Helmert}}, \code{\link[car:Contrasts]{contr.Sum}}, \code{\link[car:Contrasts]{contr.Treatment}}}

  \item{stp25settings}{\code{\link[stp25settings:set_lattice]{bw_theme}}, \code{\link[stp25settings]{farbe}}, \code{\link[stp25settings:set_opt]{get_opt}}, \code{\link[stp25settings:set_lattice]{ggplot_theme}}, \code{\link[stp25settings]{set_opt}}, \code{\link[stp25settings:set_lattice]{standard_theme}}, \code{\link[stp25settings]{which_output}}}

  \item{stp25tools}{\code{\link[stp25tools]{wrap_string}}}
}}

