% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_mean}
\alias{render_mean}
\title{render_mean
Formatiere von Zahlen nach dem APA-Style ( American Psychological Association )}
\description{
render_mean
Formatiere von Zahlen nach dem APA-Style ( American Psychological Association )
}
\examples{

rndr_median(rnorm(5), rnorm(5), 2)
rndr_median_quant(quantile(rnorm(10)), 2)
p<- c(.1259, .01, .001, .0001)
rndr_P(p)



rndr_percent(c(.2568, 99, 0.02568), c(4, 569, 25), digits = 1)
rndr_percent(2.3, 5, digits = 1)
rndr_percent(2.3, 5, digits = 1, style = 2)
rndr_percent(2.3, 5, digits = 1, style = 3)
rndr_percent(2.3, 5, digits = 1, style = 4)

}
