% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{stp25stat2}
\alias{stp25stat2}
\alias{stp25stat2-package}
\title{Analytics}
\description{
Collection of convenient functions for common statistical computations.
}
\author{
Wolfgang Peter (w.peter@statistik-peter.at)
}
