% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{Summarise}
\alias{Summarise}
\title{Summarise Data}
\usage{
Summarise(
  ...,
  fun = function(x) length(na.omit(x)),
  key = "variable",
  value = "value",
  na.action = na.pass,
  formula = NULL,
  include.total = FALSE,
  margins = include.total,
  margins_name = "Total",
  include.label = TRUE
)
}
\arguments{
\item{...}{An Long() daten Formeln, Variablen-Namen}

\item{fun}{function(x) length(na.omit(x)) an aggregate}

\item{key, value, include.label}{an Long}

\item{na.action}{an aggregate}

\item{formula}{Zeilen/Spalten Lang/Weit}

\item{include.total, margins, margins_name}{Gesamt nur bei Gruppen sinnvoll}
}
\value{
data.frame, tibble
}
\description{
Summarise: ist eine  Erweiterung  aggregate()
}
\examples{

df <- data.frame(
  month = rep(1:3, 2),
  student = rep(c("Amy", "Bob"), each = 3),
  A = c(9, 7, 6, 8, 6, 9),
  B = c(6, 7, 8, 5, 6, 7)
)

Summarise(A + B ~ student, df)

\donttest{
data(hyper, package = "stp25data")

mean3 <- function(x)
  round(mean(x, na.rm = TRUE), 2)

Summarise(
  df,
  A, B,
  by = ~ month,
  fun = mean3,
  formula = month ~ variable,
  margins = TRUE
)


Summarise(
  chol0 + chol1 + chol6 + chol12 ~ g,
  hyper,
  fun = mean,
  key = "Zeit",
  value = "Cholesterin"
)

Summarise(
  chol0 + chol1 + chol6 + chol12 ~ g,
  hyper,
  fun = mean,
  formula = variable ~ g
)

Summarise(
  chol0 + chol1 + chol6 + chol12 ~ g,
  hyper,
  fun = function(x)
    c(m = mean(x), sd = sd(x)),
  magin = TRUE,
  formula = variable ~ g
)

hyper |>
  Summarise(
    chol0, chol1, chol6, chol12,
    by = ~ g + med,
    fun = function(x)
      c(m = mean(x), sd = sd(x)),
    formula = ~ g + med
  )

  }

}
