% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{Tbll}
\alias{Tbll}
\title{Turn an object into a table}
\usage{
Tbll(...)
}
\arguments{
\item{...}{alles an Tbll}

\item{x}{object}

\item{digits}{decimal places}
}
\value{
data.frame oder liste mit data.frames
}
\description{
Turn an object into a table
}
\examples{

lm1 <- lm(breaks ~ wool + tension, data = warpbreaks)
fm1 <- aov(breaks ~ wool + tension, data = warpbreaks)
tk1 <- TukeyHSD(fm1, "tension", ordered = TRUE)


Tbll(breaks ~ wool + tension, data = warpbreaks )

Tbll(warpbreaks,
     breaks,
     wool,
     tension)
Tbll(warpbreaks,
     breaks ~ wool + tension)

Tbll(
  lm1,
  include.p = FALSE,
  include.ci = TRUE
)

}
