% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experimentelle-funktionen.R
\name{Tbll_anova}
\alias{Tbll_anova}
\title{Anova}
\usage{
Tbll_anova(..., names = NULL, include.eta = TRUE, type = c("II", "III", 2, 3))
}
\arguments{
\item{...}{lm-Objects}

\item{names}{Response}

\item{include.eta}{an tbll_extract}

\item{type}{an car::Anova}
}
\value{
tibble, data.frame
}
\description{
Anova from car::Anova
}
\examples{

ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)
lm.D9 <- lm(weight ~ group)
lm.D90 <- lm(weight ~ group - 1) # omitting intercept


Tbll_anova(lm.D9, lm.D90)


}
