% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tbll_contrasts.R
\name{Tbll_contrasts}
\alias{Tbll_contrasts}
\title{Estimate Marginal Contrasts}
\usage{
Tbll_contrasts(
  x,
  contrast = NULL,
  p_adjust = "holm",
  method = "pairwise",
  digits = 2,
  include.label = FALSE,
  ...
)
}
\arguments{
\item{x}{A statistical model. or class estimate_contrasts}

\item{contrast}{If the contrasts are not null, the contrasts are passed to the estimate_contrasts() function (character vector containing the name of the variable(s)).
If null, the name of the variable is extracted and the contrasts are evaluated separately.}

\item{p_adjust}{Can be one of "holm" (default), "tukey", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr" or "none". See the orginal estimate_contrasts()}

\item{method}{Contrast method. "pairwise"}

\item{digits}{only for Coefficient and CIs}

\item{include.label}{not working}

\item{...}{Other arguments passed to estimate_contrasts() by = modulate it}
}
\value{
data.frame
}
\description{
Stolen from modelbased::estimate_contrasts()
}
\examples{

fit <- lm(breaks ~ wool * tension, data = warpbreaks)

modelbased::estimate_contrasts(fit, contrast =  "tension") |> Tbll_contrasts()
Tbll_contrasts(fit, contrast =  "tension")
Tbll_contrasts(fit)
# require(effects)
# # Results may be misleading due to involvement in interactions
# mod <- lm(prestige ~ type * (education + income) + women, Prestige)
# Tbll_contrasts(mod)

}
