% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll_icc.R
\name{Tbll_icc}
\alias{Tbll_icc}
\alias{Tbll_icc.lmerMod}
\alias{Tbll_icc.default}
\title{Intraclass Correlation Coefficient (ICC)}
\usage{
Tbll_icc(x, ...)

\method{Tbll_icc}{lmerMod}(
  ...,
  names = NULL,
  digits = 2,
  by_group = FALSE,
  tolerance = 1e-05,
  ci = NULL,
  iterations = 100
)

\method{Tbll_icc}{default}(...)
}
\arguments{
\item{...}{lmer - fits}

\item{names, digits}{an prepare data}

\item{by_group, tolerance, ci, iterations}{geht an icc {performance}}
}
\value{
data.frame
}
\description{
stolen from performance::icc
}
\examples{
 require(lme4)

 fm1 <- lme4::lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
 Tbll_icc(fm1)

}
