% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll_likert.R
\name{Tbll_likert}
\alias{Tbll_likert}
\alias{Tbll_likert.default}
\alias{Tbll_likert.likert}
\alias{Likert}
\alias{print.likert}
\title{Tbll_likert}
\usage{
Tbll_likert(...)

\method{Tbll_likert}{default}(
  ...,
  ReferenceZero = NULL,
  include.mean = TRUE,
  include.n = FALSE,
  include.na = FALSE,
  include.order = TRUE,
  include.percent = TRUE,
  include.count = TRUE,
  decreasing = TRUE,
  labels = c("low", "neutral", "high"),
  reverse.levels = FALSE,
  reorder.levels = NA
)

\method{Tbll_likert}{likert}(
  x,
  ReferenceZero = NULL,
  include.mean = TRUE,
  include.n = FALSE,
  include.na = FALSE,
  include.order = TRUE,
  include.percent = TRUE,
  include.count = TRUE,
  labels = c("low", "neutral", "high"),
  decreasing = TRUE
)

Likert(..., labels = NULL, reverse.levels = FALSE, reorder.levels = NA)

\method{print}{likert}(x, ...)
}
\arguments{
\item{...}{Likert-Objekt oder data+Formula}

\item{ReferenceZero, labels}{ReferenceZero=2 Neutrales Element in Kombination mit
labels = c("low", "neutral", "high")}

\item{include.mean, include.n, include.na, include.order, include.percent, include.count}{Zusatz}

\item{reverse.levels}{an Likert}

\item{reorder.levels}{character ore numbers}

\item{caption, note}{an prepare Data}
}
\value{
data.frame mit attributen fuer plotlikert
}
\description{
Tbll_likert
}
\examples{

#'set.seed(1)
n <- 100
lvs <- c("--", "-", "o", "+", "++")
DF2 <- data.frame(
  Magazines = cut(rnorm(n), 5, lvs),
  Comic.books = cut(rnorm(n), 5, lvs),
  Fiction = cut(rnorm(n), 5, lvs),
  Newspapers = cut(rnorm(n), 5, lvs),
  Geschlecht = cut(rnorm(n), 2, c("m", "f"))
)

#x<-Tbll_likert(DF2, Magazines, Comic.books, Fiction, Newspapers, ReferenceZero=2)

# stp25plot::likertplot(Item   ~ . , data = x)

#
Tbll_likert(DF2, Magazines, Comic.books, Fiction, Newspapers,
            by=~Geschlecht,
            include.order=TRUE)

}
