% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll_loglm.R
\name{Tbll_loglm}
\alias{Tbll_loglm}
\title{Format Log-Linear Models}
\usage{
Tbll_loglm(
  ...,
  names = NULL,
  include.likelihood = TRUE,
  include.pearson = FALSE
)
}
\arguments{
\item{...}{MASS-loglm Objekt}

\item{names}{Modellname}

\item{include.likelihood}{logical. Likelihood Ratio  LL}

\item{include.pearson}{logical. Pearson X-Sqr}
}
\value{
data.frame
}
\description{
Format Log-Linear Models
}
\examples{

require(MASS)
Tbll_loglm(HairEyeColor, include.pearson = TRUE)


#require(stp25stat2)
Plank <- data.frame(
  therapy = c("H+A", "H+A+P", "H+A", "H+A+P", "H+A", "H+A+P", "H+A", "H+A+P"),
  mace = c("yes", "yes", "no", "no", "yes", "yes", "no", "no"),
  infarction = c("NSTEMI", "NSTEMI", "NSTEMI", "NSTEMI", "STEMI", "STEMI", "STEMI", "STEMI"),
  freq = c(7, 1, 140, 54, 63, 14, 221, 126)
)
Plank$mace_bin <- ifelse(Plank$mace == "yes", 1, 0)


# Log-Linear Models
tbl <- xtabs(freq ~ therapy + mace + infarction, data = Plank)

#  M1 	[Therapy][MACE][Infarct]
#  vollständige Unabhängigkeit aller drei Variablen
mod1 <- loglm(~ therapy + mace + infarction, data = tbl)
#  M2 	[Therapy:MACE][Infarct]
#  Therapy und MACE assoziiert, Infarkttyp unabhängig
mod2 <- loglm(~ therapy + mace + infarction + therapy:mace, data = tbl)
#  M3 	[Therapy:MACE][Infarct:MACE]
#  Therapy und Infarkttyp wirken separat über MACE, aber sind untereinander unabhängig
mod3 <- loglm(~ therapy + mace + infarction + therapy:mace + infarction:mace, data = tbl)
#  M4 	[Therapy:MACE][Infarct:MACE] [Therapy:Infarct]
#  alle Zwei-Wege alle paarweisen Abhängigkeiten, keine Dreifachinteraktion
mod4 <- loglm(~ therapy + mace + infarction + therapy:mace +
                therapy:infarction + infarction:mace, data = tbl)
 #  M5  [Therapy:MACE:Infarct]
 #  Saturated 	perfekter Fit, keine Restriktionen
mod5 <- loglm(~ therapy * mace * infarction, data = tbl)


anova(mod1, mod2, mod3, mod4) |> Tbll_loglm()
# Die ANOVA Prüft, ob ein Model signifikant besser ist als das Null-Model (M1).
# M3 ist signifikant (p < 0.001) -> M3 erklärt mehr Varianz als M1
#
# Je kleiner die Deviance umso besser ist das Modell - M3 ist hier recht gut.
Tbll_loglm(mod3)
# Hier ist ein nicht-signifikanter p-Wert besser da die Hypothese einen Modell
# testet


glm_logistic <- glm(
  mace_bin ~ therapy + infarction,
  weights = freq,
  family = binomial(),
  data = Plank
)


## Poisson-Regression (loglinear)
glm_poisson <- glm(
  freq ~ therapy * mace + infarction * mace,
  family = poisson(),
  data = Plank
)

# Odds Ratio (OR
OR <- round(exp(coef(glm_logistic))[2:3], 2)
# Risk Ratio (RR)
RR <- round(exp(coef(glm_poisson))[5:6], 2)
cbind(OR, RR)
# MACE-no-Rate war um 57\% geringer unter H+A+P2Y12 im Vergleich zu H+A.
# (Die Prozent werden aus dem RR berechnet => 1 - 0.43 = .57)

}
