% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_pca.R
\name{Tbll_pca}
\alias{Tbll_pca}
\alias{Tbll_pca_loadings}
\alias{Tbll_pca_eigen}
\alias{Tbll_pca_test}
\title{Tabelle PCA}
\usage{
Tbll_pca(x, cut = 0.3, sort = TRUE, digits = 2, ...)

Tbll_pca_loadings(x, cut = 0.3, sort = TRUE, digits = 2, ...)

Tbll_pca_eigen(x, digits = 2, extprn = NULL, ...)

Tbll_pca_test(x, ...)
}
\arguments{
\item{x}{psych::principal}

\item{cut}{display loadings}

\item{sort}{by value}

\item{digits}{digits}

\item{...}{not used}
}
\value{
data.frame or list with data.frame
}
\description{
Bortz Schuster 2010 Seite 396
}
\details{
Stichprobengroesse: Wenn je Faktor zehn oder mehr Variablken anfallen,
 ist ein n von 150 ausreichend.

 Mindestens vier Variablen muessen Ladungen > 0.60 aufweisen, oder 10 bis 12
 um 0.40 dann ist eine generalisierende Interpretation zulässig.

 Faktorladung ~ Korelation zwichen Variable und Faktor

 Kommunalität ~ Varianz die durch die Faktoren aufgeklaert wird

 Eigenwert ~ Gesamt-Varianz aller Variablen bezogen auf die Faktoren.
}
\examples{

require(magrittr)
library(stp25data)


rslt <- psych::principal(fkv, 5)

rslt  \%>\%  Tbll_pca_loadings()
rslt  \%>\%  Tbll_pca_eigen()
rslt  \%>\%  Tbll_pca_test()

}
