% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll_ranking.R
\name{Tbll_ranking}
\alias{Tbll_ranking}
\alias{Rangreihe_default}
\title{Tbll_ranking}
\usage{
Tbll_ranking(
  ...,
  na.action = na.pass,
  include.percent = TRUE,
  include.freq = TRUE,
  include.mean = TRUE,
  include.z = TRUE,
  include.na = TRUE,
  include.order = TRUE,
  decreasing = TRUE,
  digits.mean = 2
)

Rangreihe_default(
  items,
  caption = "",
  note = "",
  include.percent = TRUE,
  include.freq = TRUE,
  include.mean = TRUE,
  include.z = TRUE,
  include.na = TRUE,
  groups = NULL,
  order = TRUE,
  decreasing = TRUE,
  digits.mean = 2,
  input = NULL,
  pattern = "____"
)
}
\arguments{
\item{...}{Weitere Argumente}

\item{include.percent, include.freq, include.mean, include.z, include.na}{was soll ausgewertet werden}

\item{digits.mean, order, decreasing}{sortierung}

\item{items}{data.frame}

\item{caption, note, output}{an stp25output}

\item{groups}{gruppen}

\item{input}{Format der Items c("ranking", "ordering"),}

\item{pattern}{intern gruppen}
}
\value{
Vector
}
\description{
Rangordnungen von Objekten koennen durch eine Transformation der Rangreihen in
Intervallskalierte Merkmale ueberfuehrt werden. Die Grundidee dieser Methode geht
auf Thurstone (1927) nach dem "Law of Categorical Judgement" zurueck.
}
\details{
Dabei werden
die kumulierten Haeufigkeiten in Normalverteilte z-Werte uebergefuehrt und aus diesen
die Intervallskalierten Markmalsauspraegungen gebildet.

Literatur: Bortz, J. & Doering, N. (2006). Forschungsmethoden und
 Evaluation fuer Human-und Sozialwissenschaftler (4. Auflage). Berlin: Springer. Seite 155
}
