% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll_reg.R
\name{Tbll_reg}
\alias{Tbll_reg}
\alias{Tbll_reg_long}
\title{Regression Table}
\usage{
Tbll_reg(
  ...,
  caption = "",
  include.label = FALSE,
  names = NULL,
  digits = NULL,
  include.b = TRUE,
  include.se = FALSE,
  include.beta = FALSE,
  include.ci = TRUE,
  include.odds = FALSE,
  include.odds.ci = FALSE,
  include.statistic = FALSE,
  conf.method = "Wald",
  include.p = FALSE,
  include.stars = if (include.p) FALSE else TRUE,
  include.param = TRUE,
  include.gof = TRUE,
  include.r = TRUE,
  include.pseudo = FALSE,
  include.aic = TRUE,
  include.bic = include.aic,
  include.loglik = FALSE,
  include.minus.LL = include.loglik,
  include.test = FALSE,
  include.rmse = TRUE,
  include.custom = NULL
)

Tbll_reg_long(
  ...,
  caption = "",
  include.label = FALSE,
  digits = NULL,
  include.b = TRUE,
  include.se = FALSE,
  include.beta = FALSE,
  include.ci = TRUE,
  include.odds = FALSE,
  include.odds.ci = FALSE,
  include.statistic = TRUE,
  conf.method = "Wald",
  include.p = FALSE,
  include.stars = if (include.p) FALSE else TRUE,
  include.r = TRUE,
  include.pseudo = FALSE,
  include.aic = TRUE,
  include.bic = include.aic,
  include.param = TRUE,
  include.gof = FALSE,
  include.loglik = FALSE,
  include.minus.LL = include.loglik,
  include.custom = NULL
)
}
\description{
Bei glm  kann mit dem Parameter conf.method Wald oder z-test für
 die Test statistik gewaelt werden.
}
\examples{

 lm1 <- lm(breaks ~ wool + tension, data = warpbreaks)
 lm2 <- lm(breaks ~ wool * tension, data = warpbreaks)


 Tbll_reg(
  lm1,
  lm2,
  include.p = FALSE,
  include.ci = TRUE,
  include.se=FALSE,
  caption = "Regression Table"
)

Tbll(lm1, lm2)

}
