% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll_reliability.R
\name{Tbll_reliability}
\alias{Tbll_reliability}
\alias{Tbll_reliability.default}
\alias{Tbll_reliability.psych}
\alias{Tbll_Alpha}
\title{Tbll_reliability}
\usage{
Tbll_reliability(...)

\method{Tbll_reliability}{default}(
  ...,
  include.label = TRUE,
  include.item_statistics = TRUE,
  include.scale_statistics = TRUE,
  include.cronbachs.alpha = TRUE,
  include.inter.item.correlation = FALSE,
  revcoded = FALSE,
  check.keys = NULL,
  keys = NULL,
  digits = 2
)

\method{Tbll_reliability}{psych}(x, digits = 2)

Tbll_Alpha(..., type = 1, names = NULL)
}
\arguments{
\item{...}{Data, Formula}

\item{include.label}{Labels}

\item{revcoded, check.keys, keys}{numeric oder logikal TRUE entspricht check.keys = TRUE}

\item{digits}{Digits}

\item{caption, note}{an prepare}

\item{include.item_statistic}{Item Statistic}

\item{include.scale_statistic}{Scale Statistic}
}
\value{
list ore data.frame
}
\description{
Cronbachs Alpha oder einfach nur α ist ein Maß für die interne Konsistenz einer Skala.
}
\details{
> .9	Exzellent

 > .8	Gut / Hoch

 > .7	Akzeptabel

 > .6	Fragwürdig

 > .5	Schlecht / Niedrig

 < .5	Inakzeptabel

Quelle http://statistikguru.de/spss/reliabilitaetsanalyse/auswerten-und-berichten-2.html

Um die interne Konsistenz zu bestimmen, wurde Cronbachs Alpha für die Subskala positiver Affekt (insgesamt zehn Fragen) berechnet. Die interne Konsistenz war hoch, mit Cronbachs Alpha = .89 für positiven Affekt.


For reliability analysis, Cronbach’s alpha was calculated to assess the internal consistency of the subscale for positive affect, which consists of ten questions. The internal consistency of the questionnaire is satisfying, with Cronbach’s alpha for positive affect = .89.
}
\examples{

dat <- data.frame(
  x = c(1, 2, 3, 4, 5, 1, 2, 3, 4, 5),
  y = c(2, 2, 3, 4, 4, 3, 4, 5, 4, 2),
  z = c(3, 5, 1, 2, 3, 2, 3, 4, 4, 5),
  w = c(4, 2, 3, 5, 2, 4, 5, 4, 1, 1)
)

Tbll_reliability(dat,
                 x,
                 y,
                 z,
                 w,
                 revcoded = TRUE,
                 include.item_statistics = FALSE)

m1 <-
  Tbll_reliability(dat,
                   x,
                   y,
                   z,
                   w,
                   revcoded = TRUE,
                   include.item_statistics = FALSE)
m2 <-  Tbll_reliability(dat, x, y, z, w)

Tbll_Alpha(m1, m2)
}
