% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll-roc.R
\name{Tbll_roc}
\alias{Tbll_roc}
\title{Tabellen aus pROC::roc}
\usage{
Tbll_roc(x, digits = 2, include.order = TRUE, ...)
}
\arguments{
\item{x}{pROC::roc Objekt}

\item{digits}{numeric.}

\item{include.order}{logical.}

\item{...}{}
}
\value{
data.frame
}
\description{
Tabellen aus pROC::roc
}
\examples{
# example code
#'
require(pROC)
n <- 200
set.seed(n)
D.ex <- rbinom(n, size = 1, prob = .5)
re_scale <- function(x, min=0, max=2){
  x <- x - min(x)
  x <- x/max(x)
  x <- x*(max - min)
  round(x + min, 2)
}
DF <- data.frame(
  Group = factor(c("Control", "Ill")[D.ex + 1]),
  Cys = re_scale(rnorm(n, mean = D.ex, sd = .65), 18,84),
  Glu = re_scale(exp(2 - (D.ex) - rnorm(n, mean = 0, sd = .5)), 23, 537),
  Arg = re_scale(rnorm(n, mean = D.ex, sd = 2), 41,216),
  Gln = re_scale(rnorm(n, mean = D.ex, sd = 1.5) + D.ex, 163,1129)
) |> stp25tools::Label(
  Arg  = "Arginine",
  Cit = "Citrulline",
  Cys = "Cysteine",
  Glu = "Glutamate",
  Gln = "Glutamine"
)


roc1 <- pROC::roc(Group ~ Glu, data = DF)
Tbll_roc(roc1)
# plotROC2(roc1)

roc.list <- pROC::roc(Group ~ Glu+Gln+Arg+Cys, data = DF)
Tbll_roc(roc.list)

#plotROC2(roc.list)
}
