% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll_surv.R
\name{Tbll_surv}
\alias{Tbll_surv}
\alias{Tbll_surv.survfit}
\alias{Tbll_surv.coxph}
\alias{Tbll_surv.formula}
\alias{Tbll_coxph_uni}
\alias{tbll_extract.survfit}
\alias{tbll_extract.summary.survfit}
\alias{tbll_extract.survdiff}
\alias{tbll_extract.coxph}
\alias{tbll_extract.summary.coxph}
\title{Table for Survival Objects}
\usage{
Tbll_surv(x, ...)

\method{Tbll_surv}{survfit}(
  ...,
  data = NULL,
  names = NULL,
  include.param = FALSE,
  include.test = TRUE
)

\method{Tbll_surv}{coxph}(..., names = NULL, include.param = FALSE, include.test = TRUE)

\method{Tbll_surv}{formula}(
  formula,
  data,
  include.ci = TRUE,
  digits = 2,
  include.test = TRUE,
  ...
)

Tbll_coxph_uni(
  x,
  ...,
  include.se = FALSE,
  include.ci = TRUE,
  include.z.test = TRUE,
  include.wald.test = FALSE,
  digits = 2
)

\method{tbll_extract}{survfit}(
  x,
  include.survival = FALSE,
  include.se = FALSE,
  include.ci = TRUE,
  digits = 2,
  ...
)

\method{tbll_extract}{summary.survfit}(x, digits = NULL, include.se = FALSE, include.ci = TRUE)

\method{tbll_extract}{survdiff}(x)

\method{tbll_extract}{coxph}(x, include.param = FALSE, include.test = TRUE)

\method{tbll_extract}{summary.coxph}(
  x,
  include.se = FALSE,
  include.ci = TRUE,
  include.test = TRUE,
  digits = 2
)
}
\arguments{
\item{x}{a formula object, with the response on the left of a ~ operator, and the terms on the right. The response must be a survival object as returned by the Surv function.}

\item{...}{an summary.surfit times, censored = FALSE, scale = 1, extend=FALSE}

\item{data}{a data.frame in which to interpret the variables named in the formula, or in the subset and the weights argument.}

\item{include.param}{Regrssionstabelle}

\item{include.test}{z-Test}

\item{include.ci}{Konfidenzintervall}

\item{include.se}{Standardfehler}

\item{include.survival}{Tabelle mit allen Ueberlebenszeiten}

\item{note}{an Output}
}
\value{
data.frame or list with data.frames

data.frame
}
\description{
Log- Rank-Test ist das Standardverfahren für einfache Gruppenvergleiche

Univariate Cox regression

 stolen from www.sthda.com/english/wiki/cox-proportional-hazards-model

 regression coefficients = the hazard (risk)

 Hazard ratios = effect size of covariates exp(coef)
}
\details{
Cox regression: Hazard das unmittelbare Risiko

  HR = 1: No effect

  HR < 1: Reduction in the hazard (good prognostic factor)

  HR > 1: Increase in Hazard (bad prognostic factor)
}
\examples{

require(magrittr)
# require(stp25stat2)
require(survival)

fit0 <- survfit(Surv(futime, fustat) ~ 1, data = ovarian)
fit1 <- survfit(Surv(futime, fustat) ~ rx, data = ovarian)
fit2 <-
  survfit(Surv(futime, fustat) ~ rx + resid.ds, data = ovarian)
#summary(fit1)
Tbll(fit0)
Tbll(fit1)
Tbll(fit2)

Tbll(fit1, include.survival = TRUE)

"Survival Curve Differences"
survdiff(Surv(futime, fustat) ~ rx, data = ovarian) \%>\% Tbll()

coxph(Surv(futime, fustat) ~ rx, data = ovarian) \%>\% Tbll()

Tbll_surv(
  fit1,
  fit2,
  data = ovarian,
  include.param = TRUE,
  include.test = TRUE
)


Tbll_surv(Surv(futime, fustat) ~ rx, data = ovarian)


#'
Tbll_coxph_uni(
  lung,
  age,
  by = ~ time + status,
  include.z.test = TRUE,
  include.wald.test = FALSE
)
coxph(Surv(time, status) ~ age, data = lung) \%>\%  summary() \%>\% Tbll()

Tbll_coxph_uni(lung, age, sex,
               ph.karno, ph.ecog, wt.loss,
               by = ~ time + status)

coxph(Surv(time, status) ~ age + sex + ph.karno + ph.ecog + wt.loss,
      data = lung) \%>\%  summary() \%>\% Tbll()


# The variables sex, age and ph.ecog have highly statistically significant
# coefficients, while the coefficient for ph.karno is not significant.
#
# age and ph.ecog have positive beta coefficients, while sex has a negative
# coefficient. Thus, older age and higher ph.ecog are associated with poorer
# survival, whereas being female (sex=2) is associated with better survival.


}
