% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa.R, R/tbll_reliability.R
\name{APA}
\alias{APA}
\alias{APA.coxph}
\alias{APA.summary.table}
\alias{APA.survdiff}
\alias{APA.NULL}
\alias{APA.default}
\alias{APA.QuadTypeIndependenceTest}
\alias{APA.ScalarIndependenceTest}
\alias{APA.numeric}
\alias{APA.factor}
\alias{APA.formula}
\alias{APA.lm}
\alias{APA.glm}
\alias{APA.htest}
\alias{APA.xtabs}
\alias{APA.table}
\alias{APA.meta}
\alias{APA.psych}
\alias{APA.psych_alpha}
\title{APA Style Text-Ausgabe}
\usage{
APA(x, ...)

\method{APA}{coxph}(x, ...)

\method{APA}{summary.table}(x, ...)

\method{APA}{survdiff}(x, ...)

\method{APA}{`NULL`}(x, ...)

\method{APA}{default}(x, ...)

\method{APA}{QuadTypeIndependenceTest}(x, ...)

\method{APA}{ScalarIndependenceTest}(x, ...)

\method{APA}{numeric}(x, ...)

\method{APA}{factor}(x, ...)

\method{APA}{formula}(x, data, exclude = NA, max_factor_length = 25, ...)

\method{APA}{lm}(x, include.r = TRUE, ...)

\method{APA}{glm}(x, ...)

\method{APA}{htest}(x, ...)

\method{APA}{xtabs}(x, ...)

\method{APA}{table}(x, ...)

\method{APA}{summary.table}(x, ...)

\method{APA}{meta}(x)

\method{APA}{psych}(x)

\method{APA}{psych_alpha}(x)
}
\arguments{
\item{x}{Objekt fit, formula usw}

\item{...}{digits, data, usw}

\item{data, exclude, max_factor_length}{bei der Verwendung von Formeln}

\item{include.r}{APA.lm: R-Squar}
}
\value{
Character Vector mit einem oder meheren Eintraegen
}
\description{
Ausgabe von APA-Style formatiertem Text.
}
\examples{

APA(mpg ~ cyl, mtcars)
APA(glm(vs ~ mpg, mtcars, family = binomial()))
APA(lm(mpg ~ drat + wt + qsec, mtcars))
APA(aov(mpg ~ drat + wt + qsec, mtcars))



a <- letters[1:3]
APA(summary(table(a, sample(a))))


# Test Survival Curve Differences
require(survival)
APA(survdiff(Surv(futime, fustat) ~ ecog.ps, data = ovarian))



require(coin)


davis <- matrix(
c(3,  6,
  2, 19),
nrow = 2, byrow = TRUE
)
davis <- as.table(davis)
## Asymptotic Pearson chi-squared test
diffusion <- data.frame(
  pd = c(0.80, 0.83, 1.89, 1.04, 1.45, 1.38, 1.91, 1.64, 0.73, 1.46,
         1.15, 0.88, 0.90, 0.74, 1.21),
  age = factor(rep(c("At term", "12-26 Weeks"), c(10, 5)))
)

## Exact Wilcoxon-Mann-Whitney test
## Hollander and Wolfe (1999, p. 111)
## (At term - 12-26 Weeks)
wt <- wilcox_test(pd ~ age, data = diffusion,
                  distribution = "exact", conf.int = TRUE)

APA(xt<-chisq_test(davis))


APA(wt)

 # require(stp25data)

 # APA( ~ rrs0 + rrs1 + g, hyper)
 # APA(chol0+chol1 ~ g, hyper) das geht nicht

 # Likelihood Ratio Test
 # require(stp25stat2)

# require(stp25data)
# data("hkarz")

# fit0 <- glm(gruppe ~ 1, hkarz, family = binomial)
# fit1 <- glm(gruppe ~ tzell + lai, hkarz, family = binomial)
# fit0 <- update(fit1, . ~ -tzell- lai)


# logLik(fit0)
# logLik(fit1)

# - 2 * (logLik(fit0) -  logLik(fit1))

# lmtest::lrtest(fit1)
# APA(fit1)
# hkarz$Lai <- factor(hkarz$lai)
# Tbll_desc(hkarz,gruppe[binomial],
 #         by = ~ Lai,
 #         include.test = TRUE)



 # T-Test
#require(coin)
#APA(coin::wilcox_test(mpg ~ factor(vs), mtcars))
APA(wilcox.test(mpg ~ vs, mtcars))
APA(t.test(mpg ~ vs, mtcars))

 TeaTasting <-
matrix(c(3, 1, 1, 3),
       nrow = 2,
       dimnames = list(Guess = c("Milk", "Tea"),
                       Truth = c("Milk", "Tea")))
APA(fisher.test(TeaTasting, alternative = "greater"))

a <- letters[1:3]
APA(summary(table(a, sample(a))))

}
