% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_format.R
\name{fix_format}
\alias{fix_format}
\alias{fix_format.default}
\alias{fix_format.list}
\title{Fix Format}
\usage{
fix_format(...)

\method{fix_format}{default}(
  x,
  digits = NULL,
  names_repair = TRUE,
  include.rownames = TRUE,
  p.value = c("Pr\\\\(\\\\>", "p value", "p.value", "pvalue"),
  se = c("Std\\\\. Error", "est.std"),
  df = c("N", "Df")
)

\method{fix_format}{list}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{digits}{Nachkomastellen}

\item{names_repair}{P-value and SE}

\item{include.rownames}{include.rownames = TRUE}

\item{p.value, se, df}{pattern}
}
\value{
data.frame
}
\description{
Fix Format
}
\examples{

\donttest{

df2 <- data.frame(
term = c("A", "B", "C", "D"),
Estimate = c(23.5, .14, 5.6, 2.9876),
df1 = c(3.3, 35., 7.8, 2.1),
df = c(3, 35, 7, 2),
N = c(33, 35, 78, 21),

F.value = c(2.73, 12.444, 14.576, 30.412),
pvalue = c(0.73, 0.044, 0.056, 0.042),   stringsAsFactors =FALSE


)

x1<-fix_format(df2)


 require(car)
mod1 <- lm(conformity ~ fcategory * partner.status, data = Moore)
mod2 <-
  glm(conformity ~ fcategory * partner.status,
      data = Moore,
      family = poisson())

Anova(mod1) |> fix_format()
summary(mod1)$coefficients |> fix_format()
Anova(mod2) |> fix_format()
summary(mod2)$coefficients |> fix_format()

}
}
