% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-model_info.R
\name{Infomation}
\alias{Infomation}
\title{Regression Infomation}
\usage{
Infomation(...)
}
\arguments{
\item{...}{Regressinsmodelle}
}
\description{
Extrahiert die Variablen namen die Labels und die Daten
}
\details{
Alternative ist die Library insight

mit zb der Funktion

model_info(m4<-Hmisc::spearman2(mpg ~ factor(vs), mtcars))

insight::model_info(m2)

insight::find_response(m4)

insight::find_predictors(m4)

aber es gehen noch nicht so viel Modelle
}
\examples{

 lm1 <- lm(breaks ~ wool + tension, data = warpbreaks)
 lm2 <- lm(breaks ~ wool * tension, data = warpbreaks)
 Infomation(lm1, lm2)

}
