% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{Mittelwert}
\alias{Mittelwert}
\alias{mean2}
\alias{sd2}
\alias{length2}
\alias{median2}
\alias{Percent2}
\alias{Mean2}
\alias{Median2}
\alias{Median_ci}
\alias{Mean_ci}
\title{Mittelwert/Prozent}
\usage{
mean2(x, na.rm = TRUE)

sd2(x, na.rm = TRUE)

length2(x, na.rm = FALSE)

median2(x, na.rm = TRUE)

Percent2(x, ...)

Mean2(x, digits = 2, ...)

Median2(x, digits = 2, ...)

Median_ci(
  x,
  conf = 0.95,
  digits = 2,
  render = if (!is.null(digits)) TRUE else FALSE
)

Mean_ci(
  x,
  conf = 0.95,
  digits = 2,
  render = if (!is.null(digits)) TRUE else FALSE
)
}
\arguments{
\item{x}{an R object.}

\item{na.rm}{logical. fehlende Werte}

\item{...}{not used}

\item{digits}{numeric.digits if render=TRUE}

\item{conf}{numeric.  default = .95  95-CI}

\item{render}{logical. Formatieren}

\item{n}{numeric. to Percent2()}

\item{exclude}{missing  to Percent2()}

\item{max_factor_length}{numeric.  to Percent2()}

\item{style}{numeric.  to Percent2()}

\item{is_true_false}{logical  to Percent2()}

\item{unit}{Einheiten}
}
\value{
numeric

character
}
\description{
mean2, sd2(), length2() median2(): Copy of the base function with na.rm = TRUE.

Percent2(): Prozent-Werte Berechnen

 style = 1    50% (27)

 style = 2    27 (50%)

 style = 3    50%

 style = 4    27

 style = 5    27/54

Mean2(), Median2(): Formatierte Ausgabe

Median_ci Confidence interval for the median
stolen from Ken Aho github.com/cran/asbio/blob/master/R/ci.median.R
}
\examples{

# numeric output
mean2(rnorm(100))
sd2(rnorm(100))

# character output
Percent2(gl(2, 8, labels = c("Control", "Treat")))


 Median_ci(1:10)
 Mean_ci(1:11)


Tbll_mean(
  breaks ~ tension + wool,
  warpbreaks,
  include.custom = Median_ci

)


}
