% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R, R/tbll_desc.R
\name{Mittelwert}
\alias{Mittelwert}
\alias{mean2}
\alias{sd2}
\alias{length2}
\alias{median2}
\alias{percent2}
\alias{calc_percent}
\alias{calc_mean}
\alias{ci_median}
\alias{ci_mean}
\title{Mean}
\usage{
mean2(x, na.rm = TRUE)

sd2(x, na.rm = TRUE)

length2(x, na.rm = FALSE)

median2(x, na.rm = TRUE, ...)

percent2(x, ...)

calc_percent(
  x,
  digits = get_opt("prozent", "digits"),
  n = length(x),
  exclude = c(NA, NaN),
  max_factor_length = 25,
  style = get_opt("prozent", "style"),
  is_true_false = FALSE,
  ...
)

calc_mean(
  x,
  digits = get_opt("mean", "digits"),
  n = length(x),
  style = get_opt("mean", "style"),
  unit = NULL
)

ci_median(
  x,
  conf = 0.95,
  digits = 2,
  render = if (!is.null(digits)) TRUE else FALSE
)

ci_mean(
  x,
  conf = 0.95,
  digits = 2,
  render = if (!is.null(digits)) TRUE else FALSE
)
}
\arguments{
\item{x}{vector}

\item{na.rm, exclude}{fehlende werte}

\item{...}{platzhalter}

\item{digits}{digits if render=TRUE}

\item{n}{laenge des vectors}

\item{max_factor_length, is_true_false}{Argumente an calc_percent}

\item{style}{lang oder Kurz}

\item{unit}{Einheiten}

\item{conf}{95% CI}

\item{render}{Formatieren}
}
\description{
Lagemasse Berechnen

Format Prozent
 style = 1    50% (27)
 style = 2    27 (50%)
 style = 3    50%
 style = 4    27
 style = 5    27/54

ci_median Confidence interval for the median
stolen from https://github.com/cran/asbio/blob/master/R/ci.median.R
author Ken Aho
}
\examples{

mean2(rnorm(100))
sd2(rnorm(100))

percent2(gl(2, 8, labels = c("Control", "Treat")))

 ci_median(1:10)
 ci_mean(1:11)


  Tbll_mean(
breaks ~ tension + wool,
warpbreaks,
include.custom = ci_median

)


}
