% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R, R/tbl.R, R/tbll_effect-emmeans.R
\name{tbll_extract.summary.aov}
\alias{tbll_extract.summary.aov}
\alias{Tbll}
\alias{tbll_extract.emm_list}
\alias{tbll_extract.emmGrid}
\alias{tbll_extract.visreg}
\title{Turn an object into a table}
\usage{
\method{tbll_extract}{summary.aov}(x, ...)

Tbll(...)

\method{tbll_extract}{emm_list}(x, include.ci = TRUE, ...)

\method{tbll_extract}{emmGrid}(x, caption = "", note = "", include.ci = TRUE, ...)

\method{tbll_extract}{visreg}(x, include.ci = TRUE, digits = 2, ...)
}
\arguments{
\item{...}{an  broom::tidy zB  adjust = "tukey"}

\item{include.ci}{an broom::tidy mit conf.int = include.ci, conf.level = 0.95,}
}
\value{
data.frame oder liste mit data.frames
}
\description{
Effect Displays
}
\examples{

lm1 <- lm(breaks ~ wool + tension, data = warpbreaks)
fm1 <- aov(breaks ~ wool + tension, data = warpbreaks)
tk1 <- TukeyHSD(fm1, "tension", ordered = TRUE)


Tbll(breaks ~ wool + tension, data = warpbreaks )

Tbll(warpbreaks,
     breaks,
     wool,
     tension)
Tbll(warpbreaks,
     breaks ~ wool + tension)

Tbll(
  lm1,
  include.p = FALSE,
  include.ci = TRUE
)


#require(stpvers)
require(emmeans)

 warp.lm <- lm(breaks ~ wool * tension, data = warpbreaks)
x1<-emmeans (warp.lm,  ~ wool | tension)
Tbll(emmeans (warp.lm,  ~ wool | tension))

x2<-emmeans (warp.lm, poly ~ tension | wool)
Tbll(emmeans (warp.lm, poly ~ tension | wool))

x3<-emmeans (warp.lm, pairwise ~ tension | wool)
Tbll(emmeans (warp.lm, pairwise ~ tension | wool))

Tbll(x3, adjust = "tukey")
Tbll(x3, adjust = "scheffe")

?summary.emmGrid

}
