% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll_corr.R, R/tbll_desc.R, R/tbll_desc_long.R,
%   R/tbll_item_analysis.R
\name{Tbll_corr}
\alias{Tbll_corr}
\alias{Tbll_desc}
\alias{Tbll_desc_multi}
\alias{Tbll_mean}
\alias{Tbll_desc_long}
\alias{Tbll_desc_item}
\title{Tbll_desc}
\usage{
Tbll_corr(
  ...,
  include.label = TRUE,
  include.mean = FALSE,
  include.n = TRUE,
  include.stars = TRUE,
  include.p = FALSE,
  cor_diagonale_up = TRUE,
  type = c("pearson", "spearman")
)

Tbll_desc(
  ...,
  include.label = TRUE,
  include.n = TRUE,
  include.nr = FALSE,
  include.total = FALSE,
  include.test = FALSE,
  include.normality.tests = FALSE,
  include.multiresponse = FALSE,
  include.custom = NULL,
  include.value = NULL,
  digits = NULL,
  use.level = 1,
  use.duplicated = FALSE
)

Tbll_desc_multi(
  ...,
  by = NULL,
  digits = 0,
  include.order = TRUE,
  exclude.last.order = FALSE,
  use.level = 1,
  include.label = TRUE,
  include.n = TRUE,
  include.nr = FALSE,
  include.total = FALSE,
  include.test = FALSE,
  include.normality.tests = FALSE,
  include.custom = NULL,
  include.value = NULL
)

Tbll_mean(
  ...,
  include.total = TRUE,
  include.custom = function(x) {
     round(c(n = length(na.omit(x)), mean = mean(x,
    na.rm = TRUE), sd = sd(x, na.rm = TRUE)), 2)
 }
)

Tbll_desc_long(
  ...,
  include.range = TRUE,
  include.n = TRUE,
  include.label = TRUE,
  include.custom = NULL,
  digits = NULL,
  abbreviate = TRUE
)

Tbll_desc_item(
  ...,
  include.label = FALSE,
  include.n = TRUE,
  include.mean = TRUE,
  include.median = FALSE,
  include.sd = TRUE,
  include.range = TRUE,
  include.shapiro = TRUE,
  include.ks = FALSE,
  include.skew = TRUE,
  include.kurtosi = include.skew,
  digits = 2
)
}
\arguments{
\item{...}{an prepare_data2}

\item{include.label}{Labels ja-nein}

\item{include.mean, include.median, include.sd}{Masszahlen}

\item{include.n}{Anzahl an gueltigen Werten}

\item{include.stars, include.p}{Tbll_corr(): P-Werte,Explizite p-Werte, Sternchen als p-Werte}

\item{cor_diagonale_up}{Tbll_corr(): Diagonale oben oder unter}

\item{type}{Tbll_corr(): enweder "pearson" oder "spearman"}

\item{include.nr, include.total, include.multiresponse}{weitere param}

\item{include.test, include.normality.tests}{Test}

\item{include.custom}{fun = function(x) Auswertung (x)}

\item{include.value}{vector oder data.frame in exact der Reihenfolge wie die meassure-variablen.}

\item{digits}{Nachkommastellen}

\item{use.level, include.order, exclude.last.order}{Tbll_desc_multi weitere einstellungen}

\item{use.duplicated}{erlaube duplikate der Messwerte}

\item{by}{Gruppenvariable}

\item{include.range}{in Tbll_desc_long()}

\item{abbreviate}{Tbll_desc_long}

\item{include.shapiro, include.ks}{Normalverteilung Tests}

\item{include.skew, include.kurtosi}{Eigenschaften}
}
\value{
data.frame
}
\description{
Simple and intuitive formula and pipe-based framework for performing basic statistical Tabels.

Tbll_desc_multi: ruft direkt Tbll_desc auf (mit unterschiedlichen Parametern).

Tbll_desc_long
}
\examples{


\donttest{
require(stp25tools)
n <- 2 * 20
e <- rnorm(n)
dat <-  Label(
  data.frame(
    a = rnorm(n) + e / 2,
    b = rnorm(n) + e,
    c = rnorm(n),
    d = rnorm(n) + e * 10,
    g = gl(2, 20, labels = c("Control", "Treat"))
  ),
  a = "Alpha",
  b = "Beta",
  c = "Gamma"
)


Tbll_corr( ~ a + b + c, dat)
Tbll_corr(a ~ c, dat)
Tbll_corr(a + b + c ~ d, dat)
Tbll_corr(a + b + c ~ d, dat, groups = ~ g)

}

n<- 100
set.seed(1)
DF <- data.frame(
  sex= gl(2, n/3, labels = c("male", "female"))[sample.int(n)],
  group = gl(2, n/3, labels = c("Control", "Treat"))[sample.int(n)],
  age= runif(n, min = 18, max = 73)

)

DF<- na.omit(DF)

set_opt(
  median = list(digits = 0, style=2),
  prozent = list(style=2, null_percent_sign =  ' . ')
)


DF |>
  Tbll_desc( sex[ratio],
             group[freq],
             age[median],
             by= ~group,
             include.total=TRUE)


Tbll_desc(
  warpbreaks,
  "H1",
  breaks,
  tension,
  by = ~ wool,
  include.total = TRUE,
  include.n = FALSE,
  include.test = TRUE
)

\donttest{
warpbreaks2 <- stp25tools::Label(warpbreaks,
                     breaks	=	"The number of breaks",
                     wool	=	"The type of wool",
                     tension	=	"The level of tension")
warpbreaks2$tension2 <- as.numeric(warpbreaks2$tension)

warpbreaks2 |>
  Tbll_desc(breaks + tension ~ wool)
warpbreaks2 |>
  Tbll_desc_long(breaks + tension ~ wool)



Tbll_desc(
  warpbreaks,
  # "H1",
  breaks,
  tension,
  by = ~ wool,
  #  include.total = TRUE,
  # include.n = FALSE,
  include.test = TRUE,
  include.value = c(breaks = "ES = 26", tension = "OR = .0256")

)

x <- Tbll_desc(
  warpbreaks,
  # "H1",
  breaks,
  tension,
  by = ~ wool,
  include.total = TRUE,
  # include.n = FALSE,
  include.test = TRUE,
  include.value = data.frame(ES = 1:2, OR = 3:4)

)


Tbll_xtabs( ~ tension + wool, warpbreaks2, include.label = FALSE)

lm1 <- lm(breaks ~ wool + tension, data = warpbreaks2)
lm2 <- lm(breaks ~ wool * tension, data = warpbreaks2)


Tbll_reg(
  lm1,
  lm2,
  include.p = FALSE,
  include.ci = TRUE,
  include.se = FALSE
)


Tbll_reg(lm1, lm2)

 }


dat <- data.frame(
  Hotel =  c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1),
  Pension = c(1, 1, 0, 0, 0, 1, 1, 1, 1, 1),
  Hostel = c(1, 1, 0, 1, 1, 0, 0, 1, 0, 1),
  Ferienwohnung = c(1, 1, 1, 1, 0, 0, 0, 1, 0, 0),
  Airbnb = c(0, 0, 0, 0, 0, 1, 0, 1, 1, 1),
  Campingplatz = c(0, 0, 0, 1, 0, 1, 1, 0, 0, 0),
  Couchsurfing =c(0,0,0,0,0, 0, 1, 0, 0, 0)
)

Tbll_desc_multi(dat,
                 Hotel,
                 Pension,
                 Hostel,
                 Ferienwohnung,
                 Airbnb,
                 Campingplatz,
                 Couchsurfing)


# from Hmisc
summary(breaks ~ tension + wool, warpbreaks)
Tbll_mean(breaks ~ tension + wool, warpbreaks)



#mtcars |> Tbll_desc(mpg, cyl,  disp,  hp, drat,
#                    wt,  qsec, vs, am, gear, carb)
mtcars |> Tbll_desc_long(
  mpg[mean, 1],
  cyl[median],
  "Hallo",
  disp[0],
  hp[0],
  drat,
  wt,
  qsec[1],
  vs[freq],
  am[freq],
  gear,
  carb,
  include.range = TRUE,
  include.n = TRUE
)

mtcars |> Tbll_desc_long(
mpg,
cyl,
drat,
wt,
gear,
carb,
# by=~gear,

include.custom = function(x, by, ...){
  x <- as.numeric(x)
  c(mean= mean(x, na.rm=TRUE), sd=sd(x))
}

)


\donttest{

mtcars$G <- factor(mtcars$vs, 0:1, c("A", "B"))

mtcars <- stp25tools::Label(
  mtcars,
  mpg	= "Miles/(US) gallon",
  cyl	= "Number of cylinders",
  disp =	"Displacement (cu.in.)",
  hp	= "Gross horsepower",
  drat =	"Rear axle ratio",
  wt	= "Weight (1000 lbs)",
  qsec =	"1/4 mile time",
  vs	= "Engine (0 = V-shaped, 1 = straight)",
  am	= "Transmission (0 = automatic, 1 = manual)",
  gear =	"Number of forward gears",
  carb =	"Number of carburetors"
)

mtcars |> Tbll_desc_long(
  mpg[mean, 1],
  cyl[median],
  "Hallo",
  disp[0],
  hp[0],
  drat,
  wt,
  qsec[1],
  vs[freq],
  am[freq],
  gear,
  carb,
  by =  ~ G,
  include.range = TRUE,
  include.n = TRUE,
  include.label=FALSE
)

}


df <- data.frame(
  A = c(1,2,3,4,5, 1,1,1,2, 1,1,1,5,5,5),
  B = c(1,2,3,4,5, 2,2,2,1, 1,1,1,5,5,5),
  C = c(1,2,3,4,5, 3,3,3,4, 1,1,1,5,5,5),
  D = c(1,2,3,5,5, 4,5,4,5, 1,1,1,5,5,5),
  E = c(1,2,3,4,5, 3,2,1,1, 1,1,1,5,5,5)
)

# Tbll_desc_item(  ~ A + B + C + D + E, df)

Tbll_reliability(~ A + B + C + D + E, df,
                 include.item_statistics = TRUE)

#  df$index <-scale( rowSums(df) )
#  Tbll_item_analysis(  A + B + C + D + E ~ index, df)

}
