% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-lda.R
\name{Tbll_lda}
\alias{Tbll_lda}
\title{Tbll_lda}
\usage{
Tbll_lda(...)
}
\arguments{
\item{...}{}

\item{x}{lda {MASS}}

\item{newdata}{Test daten wenn Null dann wird model.frame(x) verwendet und an MASS::predict.lda weitergegeben.}

\item{digits}{digits}

\item{include.means}{aritmetic means}

\item{include.scal}{scaling (LD)}

\item{include.cTab}{Accuracy Table}

\item{include.svd}{Eigenvalue}
}
\value{
list with means, scal, cTab, include.svd and predictiv-Data
}
\description{
Linear Discriminant Analysis
}
\examples{

# Linear Discriminant Analysis
 require(MASS)
}
