% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll_lda.R
\name{Tbll_lda}
\alias{Tbll_lda}
\title{Tbll_lda}
\usage{
Tbll_lda(..., include.label = NULL)
}
\arguments{
\item{...}{tbll_extract.lda:
x lda {MASS}
newdata  Test daten wenn Null dann wird model.frame(x)
verwendet und an MASS::predict.lda weitergegeben.
include.svd Eigenvalue
include.cTab Accuracy Table
include.scal scaling (LD)
include.means aritmetic means
digits digits}

\item{include.label}{character. Beschriftung auf Labels umstellen}
}
\value{
list with means, scal, cTab, include.svd and predictiv-Data
}
\description{
Linear Discriminant Analysis
}
\examples{

# Linear Discriminant Analysis
 require(MASS)
}
