% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll_likert.R
\name{Tbll_likert}
\alias{Tbll_likert}
\alias{Tbll_likert.default}
\alias{Tbll_likert.likert}
\alias{Likert}
\alias{print.likert}
\title{Tbll_likert}
\usage{
Tbll_likert(...)

\method{Tbll_likert}{default}(
  ...,
  include.reference = NULL,
  include.mean = TRUE,
  include.n = FALSE,
  include.na = FALSE,
  include.order = FALSE,
  include.percent = TRUE,
  include.count = TRUE,
  include.total = FALSE,
  decreasing = TRUE,
  ReferenceZero = include.reference,
  labels = c("low", "neutral", "high"),
  reverse.levels = FALSE,
  reorder.levels = NA
)

\method{Tbll_likert}{likert}(
  x,
  include.reference = NULL,
  include.mean = TRUE,
  include.n = FALSE,
  include.na = FALSE,
  include.order = FALSE,
  include.percent = TRUE,
  include.count = TRUE,
  ReferenceZero = include.reference,
  labels = c("low", "neutral", "high"),
  decreasing = TRUE
)

Likert(
  ...,
  labels = NULL,
  reverse.levels = FALSE,
  reorder.levels = NA,
  include.total = FALSE
)

\method{print}{likert}(x, ...)
}
\arguments{
\item{...}{Likert-Objekt oder data+Formula}

\item{include.reference, labels, ReferenceZero}{numeric include.reference = 2 (drei Gruppen)
include.reference = 2.5 (zwei Gruppen)
 Neutrales Element in Kombination mit
 labels = c("low", "neutral", "high")}

\item{include.mean, include.n, include.na}{Zusatz Ergebnisse}

\item{include.order, decreasing}{sortierung nach mittelwert}

\item{include.percent, include.count}{Format Prozent/Anzahl}

\item{include.total}{logical oder string zB. include.total ="Alle"}

\item{reverse.levels}{logical  rev(item)}

\item{reorder.levels}{integer factor(item, levels(item)[reorder.levels])}

\item{x}{Likert - Objekt}
}
\value{
data.frame mit attributen fuer plotlikert

liste mit  results = data, sowie m, sd, n
}
\description{
Likert: Auszählen der Häufigkeiten
}
\examples{

#'set.seed(1)
n <- 100
lvs <- c("--", "-", "o", "+", "++")
DF2 <- data.frame(
  Magazines = cut(rnorm(n), 5, lvs),
  Comic.books = cut(rnorm(n), 5, lvs),
  Fiction = cut(rnorm(n), 5, lvs),
  Newspapers = cut(rnorm(n), 5, lvs),
  Geschlecht = cut(rnorm(n), 2, c("m", "f"))
)

#x<-Tbll_likert(DF2, Magazines, Comic.books, Fiction, Newspapers, ReferenceZero=2)

# stp25plot::likertplot(Item   ~ . , data = x)

#
Tbll_likert(DF2, Magazines, Comic.books, Fiction, Newspapers,
            by=~Geschlecht,
            include.order=TRUE)

}
