% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll_reg.R
\name{Tbll_reg}
\alias{Tbll_reg}
\alias{Tbll_reg_long}
\alias{Tbll_reg_wide}
\title{Regression Table}
\usage{
Tbll_reg(
  ...,
  include.label = FALSE,
  names = NULL,
  digits = NULL,
  include.b = TRUE,
  include.se = FALSE,
  include.beta = FALSE,
  include.ci = TRUE,
  include.odds = FALSE,
  include.odds.ci = FALSE,
  include.statistic = FALSE,
  conf.method = "Wald",
  include.p = FALSE,
  include.stars = if (include.p) FALSE else TRUE,
  include.param = TRUE,
  include.gof = TRUE,
  include.r = TRUE,
  include.pseudo = FALSE,
  include.aic = TRUE,
  include.bic = include.aic,
  include.loglik = FALSE,
  include.minus.LL = include.loglik,
  include.test = FALSE,
  include.rmse = TRUE,
  include.custom = NULL
)

Tbll_reg_long(
  ...,
  include.label = FALSE,
  include.intercept = FALSE,
  digits = NULL,
  include.b = TRUE,
  include.se = FALSE,
  include.beta = FALSE,
  include.ci = TRUE,
  include.odds = FALSE,
  include.odds.ci = FALSE,
  include.statistic = FALSE,
  conf.method = "Wald",
  include.p = FALSE,
  include.stars = if (include.p) FALSE else TRUE,
  include.r = TRUE,
  include.pseudo = FALSE,
  include.aic = FALSE,
  include.bic = include.aic,
  include.test = FALSE,
  include.devianze = FALSE,
  include.param = TRUE,
  include.rmse = FALSE,
  include.sigma = FALSE,
  include.obs = TRUE,
  include.gof = include.r | include.pseudo | include.aic | include.bic | include.test,
  include.loglik = FALSE,
  include.minus.LL = include.loglik,
  include.custom = NULL
)

Tbll_reg_wide(
  ...,
  include.label = FALSE,
  digits = NULL,
  include.b = TRUE,
  include.se = FALSE,
  include.beta = FALSE,
  include.ci = TRUE,
  include.odds = FALSE,
  include.odds.ci = FALSE,
  include.statistic = TRUE,
  conf.method = "Wald",
  include.p = FALSE,
  include.stars = if (include.p) FALSE else TRUE,
  include.r = TRUE,
  include.pseudo = FALSE,
  include.aic = TRUE,
  include.bic = include.aic,
  include.param = TRUE,
  include.gof = FALSE,
  include.loglik = FALSE,
  include.minus.LL = include.loglik,
  include.custom = NULL
)
}
\arguments{
\item{...}{ein oder mehrer Modelle}

\item{include.label}{Namen oder Labels FALSE}

\item{names}{Spaltenname}

\item{digits}{Nachkomastellen}

\item{include.b, include.ci}{estimate}

\item{include.se}{SE}

\item{include.beta}{st. Beta}

\item{include.odds, include.odds.ci}{Odds}

\item{include.statistic}{Test Statistik}

\item{conf.method}{fuer glm zb "Wald"}

\item{include.p, include.stars}{P-Werte,Sternchen}

\item{include.param}{Ausgabe der Parameter}

\item{include.gof}{ModellGuete ausgeben}

\item{include.r, include.pseudo}{R-Quadrat}

\item{include.aic, include.bic}{Guetemasse}

\item{include.loglik, include.minus.LL}{fuer glm}

\item{include.test}{Anova Test fuer das gesamt Modell}

\item{include.rmse}{RMSE}

\item{include.custom}{Eigene Spalte}
}
\description{
Bei glm  kann mit dem Parameter conf.method Wald oder z-test fuer
die Test statistik gewaelt werden.
}
\details{
RMSE: The lower the RMSE, the better a given model is able to “fit” a dataset.

AIC and  BIC  are for Selecting Regression Models
}
\examples{

 lm1 <- lm(breaks ~ wool + tension, data = warpbreaks)
 lm2 <- lm(breaks ~ wool * tension, data = warpbreaks)

 Tbll_reg(
  lm1, lm2,
  include.p = FALSE,
  include.ci = TRUE,
  include.se=FALSE
)


}
